math-skills

+++ functions/standDev.go
@@ -2,9 +2,9 @@ package functions
 
 import "math"
 
-func StandDev(nums []float64) int {
+func StandDev(nums []float64) float64 {
 	variance := float64(Variance(nums))
 	square := math.Sqrt(variance)
 	standDev := math.Round(square)
-	return int(standDev)
+	return standDev
 }
